Allow simple `npm install; npm run watch` to do everything, without sudo.
It's true the recommendation is to install bower and grunt-cli globally,
but firepad already has grunt* in devDependencies, and the local install
won't interfere with using the global commands.

And some people don't have root so would resort to a local install anyway;
the scripts make it easy to use `npm run` without bothering with the
`node_modules/.bin/` paths.